Changed structure of codebase
In this commit I'm changing the complete structure of codebase not there will be a main scraping function which will call a main function where all the scraper are grouped together. This will allow user to create more scraper according to their need. Additionally, I've also made some changes in utils so that every function will be on the correct position.

Also I've made the code more scalable, manageable, and readable.